% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blobs}
\alias{blobs}
\alias{circles}
\title{Toy datasets}
\format{
\subsection{\code{blobs}}{

A data frame with 500 rows and 2 columns:
\describe{
\item{X}{X coordinates}
\item{Y}{Y coordinates}
}
}

\subsection{\code{circles}}{

A data frame with 1,000 rows and 2 columns:
\describe{
\item{X}{X coordinates}
\item{Y}{Y coordinates}
}
}
}
\usage{
blobs

circles
}
\description{
There are two toy datasets provided in this package: \code{blobs} and \code{circles}. \code{blobs} have points
grouped into two different clusters with approximately globular shape, while \code{circles} are grouped
into two clusters shaped like a hollow circle with one being inside the other. They are used by the
demonstration functions \code{demo_blobs()} and \code{demo_circles()} respectively, as well as being accessible
in general through \code{data(blobs)} and \code{data(circles)}.

Both datasets are generated in Python using \code{scikit-learn}'s \code{make_blobs()} ("make_blobs" 2023) and
\code{make_circles()} ("make_circles" 2023).
}
\references{
“Make_blobs.” 2023. scikit-learn.
https://scikit-learn.org/stable/modules/generated/sklearn.datasets.make_blobs.html.

“Make_circles.” 2023. scikit-learn.
https://scikit-learn.org/stable/modules/generated/sklearn.datasets.make_circles.html.
}
\keyword{datasets}
