% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\alias{is_core_point}
\title{Density-based spatial clustering of applications with noise (DBSCAN)}
\usage{
dbscan(
  data,
  eps,
  min_pts,
  metric = "euclidean",
  normalise = TRUE,
  border_pts = TRUE
)

is_core_point(data, eps, min_pts, metric = "euclidean", normalise = TRUE)
}
\arguments{
\item{data}{Dataset or distance matrix to be clustered.}

\item{eps}{Maximum distance in which to find \code{min_pts} neighbours to form a cluster.}

\item{min_pts}{Minimum number of points required to be found within \code{eps} distance to form a cluster.}

\item{metric}{Distance metric. Defaults to \code{euclidean}, other options are \code{manhattan} and
\code{precomputed}. If \code{precomputed}, \code{data} must be a distance matrix.}

\item{normalise}{Logical; whether to normalise \code{data} before fitting. Defaults to \code{TRUE}.}

\item{border_pts}{Logical; whether "border points" should be included in a cluster.
Defaults to \code{TRUE}.}
}
\value{
\code{dbscan()} returns a list object of class \code{dbscan} with the following components:

\item{dataset }{ Original dataset being clustered.}
\item{eps }{ Value of \code{eps} parameter used.}
\item{min_pts }{ Value of \code{min_pts} parameter used.}
\item{metric }{ Distance metric used.}
\item{border_pts }{ Whether border points are clustered.}
\item{cluster_labs }{ Integer vector with indicating cluster membership. Noise points are assigned 0.}
\item{fitting_time }{ Total algorithm fitting time in seconds.}

\code{is_core_point()} returns a logical vector indicating whether each observation in the dataset would be
considered a core point under the given \code{eps} and \code{min_pts} parameters.
}
\description{
Applies the DBSCAN algorithm to a given dataset.
}
\details{
The most critical parameters to set are \code{min_pts} and \code{eps}.

If dataset is not yet normalised, it is highly recommended to set \code{normalise} to true, or normalise
the data first separately. As a distance-based algorithm, DBSCAN is sensitive to features being on
varying scales.

\code{border_pts=TRUE} means border points - points that do not have sufficient number of neighbours by itself
but are neighbours of a core point - are included in a cluster, and is consistent to the original algorithm
as described by Ester et al. (1996). While \code{FALSE} excludes border points, which is more consistent to the
definition of a cluster having to always have consistent minimum density and is equivalent to hierarchical
DBSCAN described in Campello et al. (2013).
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
dbscan(blobs, 5, 0.2)
}
}
\references{
Ester, Martin, Hans-Peter Kriegel, Jörg Sander, and Xiaowei Xu. 1996. “A Density-Based Algorithm for
Discovering Clusters in Large Spatial Databases with Noise.” In \emph{KDD ’96: Proceedings of the 2nd
International Conference on Knowledge Discovery and Data Mining}, 226–31. Palo Alto, CA, United States:
AAAI Press.

Campello, Ricardo J. G. B., Davoud Moulavi and Jörg Sander. 2013. "Density-Based Clustering Based on
Hierarchical Density Estimates." In \emph{PAKDD 2013: Proceedings of the 17th Pacific-Asia Conference on
Knowledge Discovery in Databases, 160-172}. Berlin, Germany: Springer.
}
\keyword{clustering}
