% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_demo.R
\name{blobs}
\alias{blobs}
\alias{circles}
\title{Demonstration of DBSCAN on toy datasets}
\usage{
blobs()

circles()
}
\description{
Applies DBSCAN to 2-D toy datasets and displays the results. \code{blobs} represents points collected into two
clearly-separated approximately-globular clusters, and \code{circles} represents points collected into two hollow
circles inside each other. Three different algorithms will be fitted into the dataset with different \code{min_pts}
and \code{eps} parameters, and scatter plots drawn to display the results to demonstrate the impact of the different
parameters.
}
\examples{
\dontrun{
blobs()
circles()
}
}
\keyword{clustering}
