% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_blobs}
\alias{demo_blobs}
\alias{demo_circles}
\title{Demonstration of DBSCAN on toy datasets}
\usage{
demo_blobs()

demo_circles()
}
\value{
There are two parts:
\itemize{
\item Printed summary table of clustering results: number of clusters generated (excluding noise),
and the four validation metrics used in \code{summary()}: connectivity, mean silhouette width, Dunn index,
and CDbw.
\item Scatter plots of the data, with generated clusters displayed as different colours.
}
}
\description{
Applies DBSCAN to 2-D toy datasets and displays the results. \code{blobs} represents points collected into
two clearly-separated approximately-globular clusters, and \code{circles} represents points collected into
two hollow  circles inside each other. Three different algorithms will be fitted into the dataset,
each with different \code{min_pts} and \code{eps} parameters, clustering summaries printed, and scatter plots
drawn to display the resulting clusters to demonstrate the impact of the different parameters.

The parameters trialled for \code{blobs} are:
\itemize{
\item \code{eps = 0.01}, \code{min_pts = 2}
\item \code{eps = 0.25}, \code{min_pts = 5}
\item \code{eps = 1}, \code{min_pts = 10}
}

The parameters trialled for \code{circles} are:
\itemize{
\item \code{eps = 0.075}, \code{min_pts = 3}
\item \code{eps = 0.1}, \code{min_pts = 5}
\item \code{eps = 0.25}, \code{min_pts = 10}
}

The other arguments are set as \code{metric = euclidean}, \code{normalise = TRUE}, and \code{border_pts = TRUE}.
}
\examples{
\dontrun{
demo_blobs()
demo_circles()
}
}
\keyword{clustering}
