% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dbscanDATA501}
\alias{plot.dbscanDATA501}
\title{Scatter plot of DBSCAN clusters}
\usage{
\method{plot}{dbscanDATA501}(x, ax1 = 1, ax2 = 2, kind = "colour", ...)
}
\arguments{
\item{x}{Object of class \code{dbscanDATA501}; will contain the dataset and cluster labels.}

\item{ax1}{Feature to be displayed on x-axis. Either column index or name. Defaults to first column.}

\item{ax2}{Feature to be displayed on y-axis. Either column index or name. Defaults to second column.}

\item{kind}{How clusters are displayed as the grouping of data points. Default is \code{colour} (i.e.
displays the clusters with different colours); other options are \code{shape} and \code{both} (colour \emph{and}
shape).}

\item{...}{Additional graphical parameters to be passed to \code{plot.default()}, or arguments for other
methods.}
}
\description{
Displays dataset on a 2-D scatter plot with the DBSCAN-generated clusters as the grouping variable.
}
\examples{
\dontrun{
data(blobs)
dbs <- dbscan(blobs, 5, 0.2)
plot(dbs)
}
}
\keyword{clustering}
