% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{Example hello function}
\usage{
rcpp_hello_world()
}
\description{
Example hello function
}
\details{
Creates a pointless two element list lorem ipsum
}
\examples{
\dontrun{
rcpp_hello_world()
}
}
