% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_scatter.R
\name{scatter.dbscan}
\alias{scatter.dbscan}
\title{Scatter plot of DBSCAN clusters.}
\usage{
scatter.dbscan(obj, x, y, how = "colour")
}
\arguments{
\item{obj}{Object of class \code{dbscan}; will contain the full dataset and cluster labels.}

\item{x}{Feature to be displayed on x-axis.}

\item{y}{Feature to be displayed on y-axis.}

\item{how}{How generated clusters are displayed as the grouping of data points. Defaults to \code{colour} (i.e.
displays the clusters with different colours); other options are \code{shape} and \code{both} (colour \emph{and} shape).}
}
\description{
Displays clustered dataset on a 2-D scatter plot with the DBSCAN-generated clusters as the grouping variable.
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
dbs <- dbscan(blobs, 5, 0.2)
scatter(dbs, x=X, y=Y, how='colour')
}
}
\keyword{clustering}
