% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot of DBSCAN clusters.}
\usage{
scatter_plot(obj, x = 1, y = 2, kind = "colour")
}
\arguments{
\item{obj}{Object of class \code{dbscan}; will contain the full dataset and cluster labels.}

\item{x}{Feature to be displayed on x-axis. Either column index or name.}

\item{y}{Feature to be displayed on y-axis. Either column index or name.}

\item{how}{How clusters are displayed as the grouping of data points. Default is \code{colour} (i.e.
displays the clusters with different colours); other options are \code{shape} and \code{both} (colour \emph{and} shape).}
}
\description{
Displays clustered dataset on a 2-D scatter plot with the DBSCAN-generated clusters as the grouping variable.
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
dbs <- dbscan(blobs, 5, 0.2)
scatter_plot(dbs, x=X, y=Y, how='colour')
}
}
\keyword{clustering}
