% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_summary.R
\name{summary.dbscan}
\alias{summary.dbscan}
\title{Summary method for DBSCAN}
\usage{
\method{summary}{dbscan}(obj)
}
\arguments{
\item{obj}{Object of class `dbscan`.}
}
\description{
Prints summary statistics for DBSCAN-generated clusters.
}
\details{
Four summary statistics: connectivity, Dunn, silhouette, CDBW
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
blobs_clust <- dbscan(blobs, 0.2, 5)
summary(blobs_clust)
}
}
\keyword{clustering}
