% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\title{Density-based spatial clustering of applications with noise (DBSCAN)}
\usage{
dbscan(
  data,
  eps,
  min_pts,
  metric = "euclidean",
  normalise = TRUE,
  borderPoints = FALSE
)
}
\arguments{
\item{data}{Dataset or distance matrix to be clustered.}

\item{eps}{Maximum distance in which to find \code{min_pts} neighbours to form a cluster.}

\item{min_pts}{Minimum number of points required to be found within \code{eps} distance to form a cluster.}

\item{metric}{Distance metric. Defaults to \code{euclidean}, other options are \code{manhattan} and
\code{precomputed}. If \code{precomputed}, \code{data} must be a distance matrix.}

\item{normalise}{Logical; whether to normalise \code{data} before fitting. Defaults to \code{TRUE}.}

\item{borderPoints}{Logical; whether "border points" should be included in a cluster.
Defaults to \code{FALSE}.}
}
\description{
Applies the DBSCAN algorithm to a given dataset.
}
\details{
\code{borderPoints = TRUE} would be equivalent to the original DBSCAN algorithm described by Ester et al.,
while \code{FALSE} would be equivalent to DBSCAN* described in Campello et al. 2013
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
dbscan(blobs, 5, 0.2)
}
}
\keyword{clustering}
