% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_summary.R
\name{summary.dbscan}
\alias{summary.dbscan}
\title{Summary method for DBSCAN}
\usage{
\method{summary}{dbscan}(obj)
}
\arguments{
\item{obj}{Object of class \code{dbscan}.}
}
\description{
Prints summary statistics for clusters generated by \code{dbscan}.
}
\details{
There are three sections to the summary:
\itemize{
\item Parameters used: \code{min_pts}, \code{eps} and \code{metric}.
\item Clustering results: total fitting time in seconds and number of clusters generated (excluding noise).
\item Clustering quality: four internal validation metrics are provided; connectivity, mean silhouette,
Dunn index, and CDbw.
}
\subsection{On quality metrics}{

Care must be taken in using connectivity, mean silhouette and Dunn index metrics to assess the quality of
clusters generated by density-based clustering
}
}
\examples{
\dontrun{
blobs <- read.csv('blobs.csv')
blobs_clust <- dbscan(blobs, 0.2, 5)
summary(blobs_clust)
}
}
\keyword{clustering}
