% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dbscanDATA501}
\alias{summary.dbscanDATA501}
\title{Summary method for DBSCAN}
\usage{
\method{summary}{dbscanDATA501}(object, ...)
}
\arguments{
\item{object}{Object of class \code{dbscanDATA501}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints summary statistics for clusters generated by \code{dbscan}.
}
\details{
There are three sections to the summary:
\itemize{
\item Parameters used in algorithm fitting.
\item Clustering results: total fitting time in seconds and number of clusters generated (excluding noise).
\item Clustering quality: results of four internal validation metrics, and a description of what should
be aimed for in terms of value:
\itemize{
\item Connectivity
\item Mean silhouette width
\item Dunn index
\item CDbw ("composed density between and within clusters")
}
}
\subsection{On quality metrics}{

Care must be taken in using connectivity, mean silhouette width and Dunn index metrics to assess the
quality of clusters generated by density-based clustering. These metrics have an underlying assumption
that the generated clusters are approximately globular in shape, whereas density-based algorithms such as
DBSCAN are capable of generating arbitrarily-shaped clusters (Moulavi et al. 2014). Only CDbw out of these
four metrics has specifically been designed for density-based clusters. And unfortunately, due to the
underlying function's limitations, CDbw would only be defined if the original clustering's distance metric
is \code{euclidean}.
}
}
\examples{
\dontrun{
data(blobs)
blobs_clust <- dbscan(blobs, 0.2, 5)
summary(blobs_clust)
}
}
\references{
Moulavi, Davoud, Pablo A. Jaskowiak, Ricardo JGB Campello, Arthur Zimek, and Jörg Sander. 2014.
"Density-based clustering validation." In \emph{Proceedings of the 2014 SIAM international conference on
data mining}, pp. 839-847. Society for Industrial and Applied Mathematics.
}
\keyword{clustering}
